+++ point_providers/LinePointProvider.h
@@ -6,12 +6,12 @@
 
 class LinePointProvider : public IPointProvider{
 public:
-    LinePointProvider(int count, int width, int height, int margin = 50);
+    LinePointProvider(int count, long width, long height, int margin = 50);
     std::vector<Point> getPoints() override;
 
 private:
     int m_count;
-    int m_width;
-    int m_height;
+    long m_width;
+    long m_height;
     int m_margin;
 };
