+++ point_providers/RandomPointProvider.h
@@ -7,14 +7,14 @@
 class RandomPointProvider : public IPointProvider {
 
 public:
-    RandomPointProvider(int count, int width, int height, int margin = 50);
+    RandomPointProvider(int count, long width, long height, int margin = 50);
 
     std::vector<Point> getPoints() override;
 
 private:
     int m_count;
-    int m_width;
-    int m_height;
+    long m_width;
+    long m_height;
     int m_margin;
 };
