+++ point_providers/SquarePointProvider.h
@@ -5,12 +5,12 @@
 
 class SquarePointProvider : public IPointProvider {
 public:
-    SquarePointProvider(int count, int width, int height, int margin = 50);
+    SquarePointProvider(int count, long width, long height, int margin = 50);
     std::vector<Point> getPoints() override;
 
 private:
     int m_count;
-    int m_width;
-    int m_height;
+    long m_width;
+    long m_height;
     int m_margin;
 };
